+++ generators/component/templates/index.js
@@ -1,5 +1,0 @@
-// @flow
-import { <%= name %>Controller, type <%= name %>ControllerProps } from './<%= name %>.controller';
-
-export type <%= name %>Props = <%= name %>ControllerProps;
-export { <%= name %>Controller as <%= name %> };